backend/requirements.txt
# Web框架
fastapi>=0.95.0
uvicorn>=0.20.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0